First working iteration
