+++ .env.example
@@ -21,10 +21,6 @@ NEXTAUTH_SECRET="secret"
 STRIPE_PUBLISHABLE_KEY=
 STRIPE_PRIVATE_KEY=
 
-# Next Auth Discord Provider
-DISCORD_CLIENT_ID=
-DISCORD_CLIENT_SECRET=
-
 # Email sending
 EMAIL_SERVER_USER=
 EMAIL_SERVER_PASSWORD=
