+++ src/app/actions/validateUserEmail.ts
@@ -1,21 +1,27 @@
 "use server"
 
 import { MembershipStatus } from "@prisma/client"
-
 import { db } from "@/services/db"
+
+import { Target } from "@/app/shared/login/types"
 
 export interface ServerActionState {
   checked: boolean
   valid: boolean
   email?: string
 }
+
 interface FormProps {
   email: string
+  target: string
 }
+
 export async function validateUserEmail(
   prevState: ServerActionState,
   data: FormProps,
 ) {
+  console.log(data)
+
   const user = await db.user.findUnique({
     where: {
       email: data.email,
@@ -28,6 +34,15 @@ export async function validateUserEmail(
       checked: true,
       valid: false,
     }
+
+  // No need for further checks if the target is the admin portal
+  if (data.target === (Target.AdminPortal as string)) {
+    return {
+      checked: true,
+      email: data.email,
+      valid: true,
+    }
+  }
 
   // User without membership means that is not valid
   const membership = await db.membership.findFirst({
