+++ src/app/shared/login/types.tsx
@@ -1,0 +1,9 @@
+export enum Target {
+  MembershipPortal = "membership_portal",
+  AdminPortal = "admin_portal",
+}
+
+export interface LoginFormProps {
+  description?: string
+  target: Target
+}
